pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quant-investor"
version = "12.0.0"
description = "single mainline 智能量化投资平台，围绕 QuantInvestor 单一主线执行结构化研究与风控"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Quant Team", email = "quant@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "TA-Lib>=0.4.28",
    "tushare>=1.2.89",
    "yfinance>=0.2.18",
    "fredapi>=0.5.0",
    "requests>=2.31.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "tqdm>=4.65.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
backtest = [
    "backtrader>=1.9.78",
]
ml = [
    "tensorflow>=2.13.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/alpha-mw/myQuant"
Repository = "https://github.com/alpha-mw/myQuant"

[project.scripts]
quant-investor = "quant_investor.cli.main:main"
app = "web.main:app"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
follow_imports = "skip"
ignore_missing_imports = true
strict_optional = false
disable_error_code = ["import-untyped", "no-any-return", "no-untyped-def", "var-annotated"]
exclude = [
    '^quant_investor/_vendor/',
    '^quant_investor/data/',
    '^quant_investor/data_manager\.py$',
    '^quant_investor/enhanced_data_layer\.py$',
    '^quant_investor/kronos_predictor\.py$',
    '^quant_investor/forecast_snapshot_store\.py$',
    '^quant_investor/reporting/',
    '^quant_investor/learning/',
    '^quant_investor/monitoring/',
    '^quant_investor/market/download.*\.py$',
    '^quant_investor/market/run_pipeline\.py$',
    '^quant_investor/fetch_.*\.py$',
    '^quant_investor/alpha.*\.py$',
    '^quant_investor/stock_database\.py$',
    '^quant_investor/news_analysis\.py$',
    '^quant_investor/macro_terminal_tushare\.py$',
    '^quant_investor/agents/subagents/',
    '^quant_investor/agents/orchestrator\.py$',
    '^quant_investor/agents/narrator_agent\.py$',
    '^quant_investor/agent_orchestrator\.py$',
    '^quant_investor/kline_backends/',
    '^quant_investor/factor_analyzer\.py$',
    '^quant_investor/intelligence\.py$',
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=quant_investor --cov-report=term-missing --cov-report=html:results/htmlcov"

[tool.coverage.run]
source = ["quant_investor"]
omit = ["*/tests/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.hatch.build]
skip-excluded-dirs = true
exclude = [
    "/.uv-cache",
    "/.uv-python",
    "/.venv",
    "/.venv-managed",
    "/venv",
    "/frontend/node_modules",
    "/**/__pycache__",
    "/**/*.pyc",
    "/**/.DS_Store",
]

[tool.hatch.build.targets.sdist]
include = [
    "/quant_investor",
    "/README.md",
    "/pyproject.toml",
    "/requirements.txt",
]

[tool.hatch.build.targets.wheel]
packages = ["quant_investor"]
